Fixed a bug in the table extension algorithm.
Thanks to G. Hohmann (g.hohmann@wiss-ki.eu) for testing data and support.

The table extension algorithm only extends the tables if more than 16M
values (s/p/o) (this means the id of the s/p/o-tables gets higher than
16 M) are loaded. However if a file has a high connectivity
there are more triples than s/p/o-constructs. Then the triple-id hits
this value first. As a consequence the import goes on, but is unable to
store more triples, because the triple id already has hit the mediumint
limit. As soon as the s/p/o tables hit the 16M limit, the tables get
extended and storing continues in the triple table. So you will not find
the problem if you just test the first triples and the last triples of a
file. However the triples in the middle might be missing.

This patch is provided by the WissKI-Project (http://www.wiss-ki.eu) -
Have a look :)