Fixed a bug with wrong order in unions
The following query did not give the right output:
SELECT ?x0 ?out WHERE { ?x0 rdf:type
<http://erlangen-crm.org/101001/E84_Information_Carrier>  . { { ?x0
<http://erlangen-crm.org/101001/P1_is_identified_by> ?x1 } UNION { ?x1
<http://erlangen-crm.org/101001/P1i_identifies> ?x0 } }  . ?x1 rdf:type
<http://erlangen-crm.org/101001/E42_Identifier> . ?x1
<http://erlangen-crm.org/101001/P3_has_note> ?out .  } ORDER BY ?out
LIMIT 25 OFFSET 50

The resulting sql statement has the following style:
( SELECT ... ORDER BY ... ) UNION ALL ( SELECT ... ORDER BY ... ) LIMIT.

Unfortunately the MYSQL reference manual
(http://dev.mysql.com/doc/refman/5.1/en/union.html) tells:
However, use of ORDER BY for individual SELECT statements implies
nothing about the order in which the rows appear in the final result
because UNION by default produces an unordered set of rows. Therefore,
the use of ORDER BY in this context is typically in conjunction with
LIMIT, so that it is used to determine the subset of the selected rows
to retrieve for the SELECT, even though it does not necessarily affect
the order of those rows in the final UNION result. If ORDER BY appears
without LIMIT in a SELECT, it is optimized away because it will have no
effect anyway.
To use an ORDER BY or LIMIT clause to sort or limit the entire UNION
result, parenthesize the individual SELECT statements and place the
ORDER BY or LIMIT after the last one.

Did that. :)