PHP 7.4 related fixes: regex (1) and array access (2) related
(1)
regex raised errors like:

  "preg_match(): Compilation failed: escape sequence is invalid in character
   class at offset 56"

I couldn't locate the character, which was invalid, therefore i replaced
the whole part with simple datetime related functions to get the same
functionality. Added 1 test to prove behavior.

(2)
In ARC2_StoreSelectQueryHandler it sometimes try to access an int-variable like
an array. Added a check, which avoids the error:

  "Trying to access array offset on value of type int"